+++ bin/repl.js
@@ -17,9 +17,9 @@ import {
   printSummary,
   printTable,
   printVulnerabilities,
-} from "../display.js";
-import { createBom } from "../index.js";
-import { validateBom } from "../validator.js";
+} from "../lib/helpers/display.js";
+import { createBom } from "../lib/cli/index.js";
+import { validateBom } from "../lib/helpers/validator.js";
 
 const options = {
   useColors: true,
