+++ lib/cli/index.js
@@ -27,7 +27,7 @@ import {
   getOriginUrl,
   gitTreeHashes,
   listFiles,
-} from "./envcontext.js";
+} from "../helpers/envcontext.js";
 import {
   CARGO_CMD,
   CLJ_CMD,
@@ -143,7 +143,7 @@ import {
   readZipEntry,
   recomputeScope,
   splitOutputByGradleProjects,
-} from "./utils.js";
+} from "../helpers/utils.js";
 let url = import.meta.url;
 if (!url.startsWith("file://")) {
   url = new URL(`file://${import.meta.url}`).toString();
@@ -155,22 +155,22 @@ const selfPJson = JSON.parse(
 );
 const _version = selfPJson.version;
 import { gte, lte } from "semver";
-import { findJSImportsExports } from "./analyzer.js";
+import { findJSImportsExports } from "../helpers/analyzer.js";
 import {
   executeOsQuery,
   getBinaryBom,
   getCargoAuditableInfo,
   getDotnetSlices,
   getOSPackages,
-} from "./binary.js";
-import { collectOSCryptoLibs } from "./cbomutils.js";
+} from "../managers/binary.js";
+import { collectOSCryptoLibs } from "../helpers/cbomutils.js";
 import {
   addSkippedSrcFiles,
   exportArchive,
   exportImage,
   getPkgPathList,
   parseImageName,
-} from "./docker.js";
+} from "../managers/docker.js";
 
 const isWin = _platform() === "win32";
