+++ lib/helpers/cbomutils.js
@@ -1,6 +1,6 @@
 import { readFileSync } from "node:fs";
 import { join } from "node:path";
-import { executeOsQuery } from "./binary.js";
+import { executeOsQuery } from "../managers/binary.js";
 import { convertOSQueryResults, dirNameStr } from "./utils.js";
 const cbomosDbQueries = JSON.parse(
   readFileSync(join(dirNameStr, "data", "cbomosdb-queries.json"), "utf-8"),
