+++ lib/helpers/utils.js
@@ -49,7 +49,7 @@ import {
 import toml from "toml";
 import { IriValidationStrategy, validateIri } from "validate-iri";
 import { xml2js } from "xml-js";
-import { getTreeWithPlugin } from "./piptree.js";
+import { getTreeWithPlugin } from "../managers/piptree.js";
 
 let url = import.meta.url;
 if (!url.startsWith("file://")) {
