+++ lib/server/server.js
@@ -7,8 +7,8 @@ import process from "node:process";
 import url from "node:url";
 import bodyParser from "body-parser";
 import connect from "connect";
-import { createBom, submitBom } from "./index.js";
-import { postProcess } from "./postgen.js";
+import { createBom, submitBom } from "../cli/index.js";
+import { postProcess } from "../stages/postgen/postgen.js";
 
 import compression from "compression";
