+++ lib/stages/postgen/postgen.js
@@ -3,7 +3,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import process from "node:process";
 import { PackageURL } from "packageurl-js";
-import { DEBUG_MODE, dirNameStr, hasAnyProjectType } from "./utils.js";
+import { DEBUG_MODE, dirNameStr, hasAnyProjectType } from "../../helpers/utils.js";
 
 /**
  * Filter and enhance BOM post generation.
