+++ lib/stages/pregen/pregen.js
@@ -8,8 +8,8 @@ import {
   installSdkmanTool,
   isNvmAvailable,
   isSdkmanAvailable,
-} from "./envcontext.js";
-import { DEBUG_MODE, hasAnyProjectType } from "./utils.js";
+} from "../../helpers/envcontext.js";
+import { DEBUG_MODE, hasAnyProjectType } from "../../helpers/utils.js";
 
 /**
  * Method to prepare the build environment for BOM generation purposes.
